# docs/swagger.yaml
basePath: /api
definitions:
...
  edutalks_internal_models.News:
    properties:
      color:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      sticker:
        type: string
      title:
        type: string
    type: object
...
  internal_handlers.createNewsRequest:
    properties:
      color:
        type: string
      content:
        type: string
      image_url:
        type: string
      sticker:
        type: string
      title:
        type: string
    type: object
...
  internal_handlers.updateNewsRequest:
    properties:
      color:
        type: string
      content:
        type: string
      image_url:
        type: string
      sticker:
        type: string
      title:
        type: string
    type: object
host: edutalks.ru
info:
  contact:
    email: support@edutalks.ru
    name: EduTalks Support
    url: https://edutalks.ru
  description: Документация API Edutalks (регистрация, логин, токены, статьи, логи
    и т.д.).
  title: Edutalks API
  version: "1.0"
paths:
...
  /api/admin/news:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные новости
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createNewsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Новость создана
          schema:
            type: string
        "400":
          description: Ошибка запроса
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Создать новость (только admin)
      tags:
      - admin-news
  /api/admin/news/{id}:
    delete:
      parameters:
      - description: ID новости
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Удалено
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Удалить новость (только admin)
      tags:
      - admin-news
    patch:
      parameters:
      - description: ID новости
        in: path
        name: id
        required: true
        type: integer
      - description: Новое содержимое
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateNewsRequest'
      responses:
        "200":
          description: Обновлено
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Обновить новость (только admin)
      tags:
      - admin-news
  /api/admin/news/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Файл изображения (jpeg/png/webp/gif)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: 'url: публичная ссылка'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Request Entity Too Large
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузка изображения для новости
      tags:
      - news
...
  /api/news:
    get:
      parameters:
      - description: Номер страницы (начиная с 1)
        in: query
        name: page
        type: integer
      - description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/edutalks_internal_models.News'
            type: array
      summary: Получить список новостей
      tags:
      - news
  /api/news/{id}:
    get:
      parameters:
      - description: ID новости
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/edutalks_internal_models.News'
        "404":
          description: Не найдено
          schema:
            type: string
      summary: Получить новость по ID
      tags:
      - news
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"